+++ CHANGELOG.md
@@ -9,6 +9,10 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ## Unreleased
 
+### Added
+
+- Allow using the `@builder` directive on fields https://github.com/nuwave/lighthouse/pull/1687
+
 ## 5.1.0
 
 ### Added